Move redacted logging to logger
