+++ src/lib/controller/store.ts
@@ -7,11 +7,10 @@ import * as ramda from "ramda";
 
 import { Capability } from "../capability";
 import { PeprStore } from "../k8s";
-import Log from "../logger";
+import Log, { redactedPatch, redactedStore } from "../logger";
 import { DataOp, DataSender, DataStore, Storage } from "../storage";
 import { fillCache, flushCache } from "./migrateCache";
 
-export const redactedValue = "**redacted**";
 const namespace = "pepr-system";
 export const debounceBackoff = 5000;
 
@@ -175,37 +174,4 @@ export class PeprControllerStore {
       Log.error(err, "Failed to create Pepr store");
     }
   };
-}
-
-export function redactedStore(store: PeprStore): PeprStore {
-  const redacted = process.env.PEPR_STORE_REDACT_VALUES === "true";
-  return {
-    ...store,
-    data: Object.keys(store.data).reduce((acc: Record<string, string>, key: string) => {
-      acc[key] = redacted ? redactedValue : store.data[key];
-      return acc;
-    }, {}),
-  };
-}
-
-export function redactedPatch(patch: Record<string, Operation> = {}): Record<string, Operation> {
-  const redacted = process.env.PEPR_STORE_REDACT_VALUES === "true";
-
-  if (!redacted) {
-    return patch;
-  }
-
-  const redactedCache: Record<string, Operation> = {};
-
-  Object.keys(patch).forEach(key => {
-    const operation = patch[key];
-    const redactedKey = key.includes(":") ? key.substring(0, key.lastIndexOf(":")) + ":**redacted**" : key;
-    const redactedOperation: Operation = {
-      ...operation,
-      ...("value" in operation ? { value: redactedValue } : {}),
-    };
-    redactedCache[redactedKey] = redactedOperation;
-  });
-
-  return redactedCache;
 }
